intelligence/requirements.txt
# Apollo Intelligence Platform - Python Dependencies
# Intelligence & OSINT Integration - Agent 3

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async & Task Processing
celery==5.3.4
redis==5.0.1
aiohttp==3.9.0
asyncio==3.4.3
httpx==0.25.1

# Database & Search
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Message Queue
kombu==5.3.4
amqp==5.2.0

# Web Scraping & Parsing
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
requests==2.31.0
urllib3==2.1.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Blockchain APIs
web3==6.11.3
bitcoin==1.1.42
eth-account==0.10.0
eth-utils==2.3.1

# Cryptography
cryptography==41.0.7
pycryptodome==3.19.0

# Network & DNS
dnspython==2.4.2
python-whois==0.8.0
pydig==0.4.0

# IP & Geolocation
geoip2==4.7.0
ipaddress==1.0.23
ipwhois==1.2.0

# Social Media APIs
tweepy==4.14.0
python-telegram-bot==20.7

# Threat Intelligence
shodan==1.31.0
censys==2.2.8

# Rate Limiting & Caching
python-redis-rate-limit==0.0.7
cachetools==5.3.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx-mock==0.7.1

# Type Checking
mypy==1.7.1
types-redis==4.6.0.11
types-requests==2.31.0.10

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
pylint==3.0.3

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14

# Security
bandit==1.7.5
safety==2.3.5
python-jose[cryptography]==3.3.0

# Performance
uvloop==0.19.0
orjson==3.9.10

# Additional OSINT Tools
phonenumbers==8.13.26
python-magic==0.4.27
pillow==10.1.0

# Facial Recognition
face-recognition>=1.3.0
dlib>=19.24.0
opencv-python>=4.8.0

# Voice Recognition - Whisper Transcription
openai-whisper>=20231117
faster-whisper>=0.10.0

# Voice Recognition - Speaker Embeddings
resemblyzer>=0.1.1.dev0
speechbrain>=0.5.16

# Voice Recognition - Speaker Diarization
pyannote-audio>=3.1.0

# Voice Recognition - Audio Processing
torchaudio>=2.1.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1
praat-parselmouth>=0.4.3

# Blockchain Specific
base58==2.1.1
ecdsa==0.18.0
mnemonic==0.21

# Network Scanning
python-nmap==0.7.1
scapy==2.5.0

# Data Validation
validators==0.22.0
jsonschema==4.20.0

# API Documentation
fastapi-swagger-ui==0.0.6

# ============== Dark Web Monitoring (Agent 5) ==============
# Tor SOCKS Proxy
aiohttp-socks>=0.8.4
PySocks>=1.7.1
stem>=1.8.2  # Tor controller

# TimescaleDB
asyncpg>=0.29.0

# Additional async utilities
aiolimiter>=1.1.0

# Regex enhancements
regex>=2023.0.0

# Telegram
telethon>=1.34.0
pyrogram>=2.0.0

# Multi-part form support
python-multipart>=0.0.6
